Run application test only if apiPackages ff is disabled
