+++ tests/ui-tests/tests/expose-api-test.js
@@ -4,19 +4,16 @@ import kymaConsole from '../commands/console';
 import common from '../commands/common';
 import address from '../utils/address';
 import { retry, retryInterval } from '../utils/retry';
-import { testPluggable } from '../setup/test-pluggable';
 import { NamespaceManager } from '../setup/namespace-manager';
-import { k8sApiDeployment, k8sApiService } from './../setup/k8s-api';
+import { k8sApiDeployment, k8sApiService } from '../setup/k8s-api';
 
 common.setRandomNamespaceName();
 let page, browser;
 let token = ''; // eslint-disable-line no-unused-vars
 
-// TODO: Move application tests to a separate file
-const REQUIRED_MODULE = 'application';
 const namespaceInstaller = new NamespaceManager(config.testNamespace);
 
-describe('Console basic tests', () => {
+describe('Expose API test', () => {
   beforeAll(async () => {
     try {
       await namespaceInstaller.createIfDoesntExist();
@@ -146,57 +143,5 @@ describe('Console basic tests', () => {
     await retryInterval(() =>
       callExposedAPI({ expectedStatusCode: 401 }, 1000, 60),
     );
-  });
-
-  test('Check if Application exist', async () => {
-    const applicationsUrl = address.console.getApplications();
-    await Promise.all([
-      page.goto(applicationsUrl),
-      page.waitForNavigation({
-        waitUntil: ['domcontentloaded', 'networkidle0'],
-      }),
-    ]);
-    const applications = await kymaConsole.getApplicationNames(page);
-    console.log('Check if application exists', applications);
-    expect(applications).not.toContain(config.testApp);
-  });
-
-  testPluggable(REQUIRED_MODULE, 'Create Application', async () => {
-    await kymaConsole.createApplication(page, config.testApp);
-    const applications = await kymaConsole.getApplicationNames(page);
-    expect(applications).toContain(config.testApp);
-  });
-
-  testPluggable(REQUIRED_MODULE, 'Go to details and back', async () => {
-    const frame = await kymaConsole.waitForConsoleCoreFrame(page);
-    await frame.waitForXPath(
-      `//a[contains(@data-e2e-id, 'application-name') and contains(string(), "${config.testApp}")]`,
-    );
-    await kymaConsole.openLinkOnFrame(
-      page,
-      '[data-e2e-id=application-name]',
-      config.testApp,
-    );
-    frame.waitForSelector('.fd-breadcrumb__link');
-    frame.click('.fd-breadcrumb__link');
-  });
-
-  testPluggable(REQUIRED_MODULE, 'Delete Application', async () => {
-    const frame = await kymaConsole.waitForConsoleCoreFrame(page);
-    await frame.waitForXPath(
-      `//a[contains(@data-e2e-id, 'application-name') and contains(string(), "${config.testApp}")]`,
-    );
-    const initialApplications = await kymaConsole.getApplicationNames(page);
-    await kymaConsole.deleteApplication(page, config.testApp);
-    const applications = await retry(async () => {
-      const applicationsAfterRemoval = await kymaConsole.getApplicationNames(
-        page,
-      );
-      if (initialApplications <= applicationsAfterRemoval) {
-        throw new Error(`Application ${config.testApp} was not yet removed`);
-      }
-      return applicationsAfterRemoval;
-    }, 5);
-    expect(applications).not.toContain(config.testApp);
   });
 });
